Protected endpoint add, SQL copied from branch Alex. Need to be tested

+++ part3/app/api/v1/reviews.py
@@ -1,5 +1,6 @@
 from flask_restx import Namespace, Resource, fields
 from app.services import facade
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 api = Namespace('reviews', description='Review operations')
 
@@ -16,6 +17,9 @@ class ReviewList(Resource):
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
+    @api.response(400, 'You cannot review your own place')
+    @api.response(400, 'You have already reviewed this place')
+    @jwt_required()
     def post(self):
         """Register a new review"""
         review_data = api.payload
@@ -26,7 +30,10 @@ class ReviewList(Resource):
         if not user:
             return {'error': 'User not found'}, 400
         if place.owner.id == user.id:
-            return {'error': 'User cannot review their own place'}, 400
+            return {'error': 'You cannot review your own place'}, 400
+        existing_review = facade.get_review_by_place_and_user(review_data['place_id'], review_data['user_id'])
+        if existing_review:
+            return {'error': 'You have already reviewed this place'}, 400
         try:
             new_review = facade.create_review(review_data)
             return new_review.to_dict(), 201
@@ -53,13 +60,17 @@ class ReviewResource(Resource):
     @api.response(200, 'Review updated successfully')
     @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
+    @api.response(403, 'Unauthorized action')
+    @jwt_required()
     def put(self, review_id):
         """Update a review's information"""
         review_data = api.payload
-        review = facade.get_review(review_id)
+        review = facade.get_review_by_id(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
-        
+        current_user_id = get_jwt_identity()
+        if review.user_id != current_user_id:
+            return {'error': 'Unauthorized action.'}, 403
         try:
             facade.update_review(review_id, review_data)
             return {'message': 'Review updated successfully'}, 200
@@ -68,12 +79,16 @@ class ReviewResource(Resource):
 
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
+    @api.response(403, 'Unauthorized action')
+    @jwt_required
     def delete(self, review_id):
         """Delete a review"""
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
-        
+        current_user_id = get_jwt_identity()
+        if review.user_id != current_user_id:
+            return {'error': 'Unauthorized action.'}, 403
         try:
             facade.delete_review(review_id)
             return {'message': 'Review deleted successfully'}, 200
